feat: add penguins & guns db refactor: utils structure fix: tsconfig.json
