+++ app-db-data/guns/guns.csv
@@ -1,0 +1,11 @@
+Название,Тайл,Тип оружия,Урон,Стоимость,Дальность
+Обрез,1g,Пистолет,50,1000,50
+Дробаш,2g,Дробовик,100,1500,30
+Пуляло,3g,Винтовка,150,2000,200
+БОЛЬШАЯ ПУШКА,4g,Пулемет,200,3000,300
+Тра-та-тало,5g,Гранатомет,300,4000,100
+маленькая пушка,6g,Пистолет,50,1000,50
+БУМ-БУМ,7g,Бомба,500,5000,1
+Пшшшшшш,8g,Снайперская винтовка,250,2500,500
+Краснознаменный пулемет имени моей бабушки,9g,Пулемет,200,3000,300
+ОГНЕМЁТ,10g,Огнемет,400,6000,20
