+++ app-db-data/guns/guns.json
@@ -1,0 +1,82 @@
+[
+  {
+    "Название": "Обрез",
+    "Тайл": "1g",
+    "Тип оружия": "Пистолет",
+    "Урон": "50",
+    "Стоимость": "1000",
+    "Дальность": "50"
+  },
+  {
+    "Название": "Дробаш",
+    "Тайл": "2g",
+    "Тип оружия": "Дробовик",
+    "Урон": "100",
+    "Стоимость": "1500",
+    "Дальность": "30"
+  },
+  {
+    "Название": "Пуляло",
+    "Тайл": "3g",
+    "Тип оружия": "Винтовка",
+    "Урон": "150",
+    "Стоимость": "2000",
+    "Дальность": "200"
+  },
+  {
+    "Название": "БОЛЬШАЯ ПУШКА",
+    "Тайл": "4g",
+    "Тип оружия": "Пулемет",
+    "Урон": "200",
+    "Стоимость": "3000",
+    "Дальность": "300"
+  },
+  {
+    "Название": "Тра-та-тало",
+    "Тайл": "5g",
+    "Тип оружия": "Гранатомет",
+    "Урон": "300",
+    "Стоимость": "4000",
+    "Дальность": "100"
+  },
+  {
+    "Название": "маленькая пушка",
+    "Тайл": "6g",
+    "Тип оружия": "Пистолет",
+    "Урон": "50",
+    "Стоимость": "1000",
+    "Дальность": "50"
+  },
+  {
+    "Название": "БУМ-БУМ",
+    "Тайл": "7g",
+    "Тип оружия": "Бомба",
+    "Урон": "500",
+    "Стоимость": "5000",
+    "Дальность": "1"
+  },
+  {
+    "Название": "Пшшшшшш",
+    "Тайл": "8g",
+    "Тип оружия": "Снайперская винтовка",
+    "Урон": "250",
+    "Стоимость": "2500",
+    "Дальность": "500"
+  },
+  {
+    "Название": "Краснознаменный пулемет имени моей бабушки",
+    "Тайл": "9g",
+    "Тип оружия": "Пулемет",
+    "Урон": "200",
+    "Стоимость": "3000",
+    "Дальность": "300"
+  },
+  {
+    "Название": "ОГНЕМЁТ",
+    "Тайл": "10g",
+    "Тип оружия": "Огнемет",
+    "Урон": "400",
+    "Стоимость": "6000",
+    "Дальность": "20"
+  }
+]
