+++ app-db-data/penguins/penguins.csv
@@ -1,0 +1,11 @@
+Название,Тайл,Стоимость,Скорость,HP,Скрытность,Радиус обзора,Роль,Шанс огня по своим,Шанс вылечить врага
+Рядовой,1c,1,1x,10,высокая,высокий,универсальный,5,0
+Ковальски,2c,7,1x,25,средняя,средний,координатор,0,0
+Рико,3c,5,"1,5x",25,очень низкая,низкий,дамаг,50,0
+Шкипер,4c,10,1x,30,средняя,средний,офицер,0,0
+BBC Documentary Robot,5c,0,1x,200,низкая,высокий,особая,0,0
+Снейк,6c,15,3x,15,очень высокая,очень высокий,разведчик,0,0
+Макс Пейн,7c,15,0.5x,100,низкая,низкий,танк,15,0
+Сорвиголова,8c,6,0.3x,70,низкая,глаукома,танк,40,0
+Миротворец,9c,12,1.2x,35,низкая,низкий,дамаг,50,0
+Буквально Эндрю Гарфилд,10c,10,0.4x,20,средняя,средний,медик,0,15
