+++ app-db-data/penguins/penguins.json
@@ -1,0 +1,122 @@
+[
+  {
+    "Название": "Рядовой",
+    "Тайл": "1c",
+    "Стоимость": 1,
+    "Скорость": "1x",
+    "HP": 10,
+    "Скрытность": "высокая",
+    "Радиус обзора": "высокий",
+    "Роль": "универсальный",
+    "Шанс огня по своим": 5,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "Ковальски",
+    "Тайл": "2c",
+    "Стоимость": 7,
+    "Скорость": "1x",
+    "HP": 25,
+    "Скрытность": "средняя",
+    "Радиус обзора": "средний",
+    "Роль": "координатор",
+    "Шанс огня по своим": 0,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "Рико",
+    "Тайл": "3c",
+    "Стоимость": 5,
+    "Скорость": "1,5x",
+    "HP": 25,
+    "Скрытность": "очень низкая",
+    "Радиус обзора": "низкий",
+    "Роль": "дамаг",
+    "Шанс огня по своим": 50,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "Шкипер",
+    "Тайл": "4c",
+    "Стоимость": 10,
+    "Скорость": "1x",
+    "HP": 30,
+    "Скрытность": "средняя",
+    "Радиус обзора": "средний",
+    "Роль": "офицер",
+    "Шанс огня по своим": 0,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "BBC Documentary Robot",
+    "Тайл": "5c",
+    "Стоимость": 0,
+    "Скорость": "1x",
+    "HP": 200,
+    "Скрытность": "низкая",
+    "Радиус обзора": "высокий",
+    "Роль": "особая",
+    "Шанс огня по своим": 0,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "Снейк",
+    "Тайл": "6c",
+    "Стоимость": 15,
+    "Скорость": "3x",
+    "HP": 15,
+    "Скрытность": "очень высокая",
+    "Радиус обзора": "очень высокий",
+    "Роль": "разведчик",
+    "Шанс огня по своим": 0,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "Макс Пейн",
+    "Тайл": "7c",
+    "Стоимость": 15,
+    "Скорость": "0.5x",
+    "HP": 100,
+    "Скрытность": "низкая",
+    "Радиус обзора": "низкий",
+    "Роль": "танк",
+    "Шанс огня по своим": 15,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "Сорвиголова",
+    "Тайл": "8c",
+    "Стоимость": 6,
+    "Скорость": "0.3x",
+    "HP": 70,
+    "Скрытность": "низкая",
+    "Радиус обзора": "глаукома",
+    "Роль": "танк",
+    "Шанс огня по своим": 40,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "Миротворец",
+    "Тайл": "9c",
+    "Стоимость": 12,
+    "Скорость": "1.2x",
+    "HP": 35,
+    "Скрытность": "низкая",
+    "Радиус обзора": "низкий",
+    "Роль": "дамаг",
+    "Шанс огня по своим": 50,
+    "Шанс вылечить врага": 0
+  },
+  {
+    "Название": "Буквально Эндрю Гарфилд",
+    "Тайл": "10c",
+    "Стоимость": 10,
+    "Скорость": "0.4x",
+    "HP": 20,
+    "Скрытность": "средняя",
+    "Радиус обзора": "средний",
+    "Роль": "медик",
+    "Шанс огня по своим": 0,
+    "Шанс вылечить врага": 15
+  }
+]
