+++ src/utils/resource-loaders/DogAnimationLoader.js
@@ -1,4 +1,4 @@
-import config from '../../assets/animations/dogs.json';
+import config from '../../../assets/animations/dogs.json';
 
 class DogAnimationLoader {
     // frame names loaded as dog_XX_attack_frame_YY without leading zeroes!
