+++ src/utils/resource-loaders/animation-loader.js
