+++ tsconfig.json
@@ -1,7 +1,14 @@
 {
     "compilerOptions": {
-        "lib": ["es2017", "dom"],
-        "target": "ES6",
+        "rootDir": ".",
+        "baseUrl": ".",
+        "paths": {
+            "src/*": [
+                "src/*"
+            ]
+        },
+        "lib": ["ES2020", "DOM"],
+        "target": "ES2020",
         "allowJs": true,
         "checkJs": true,
         "outDir": "docs/",
